backend/vitals-generator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
elasticsearch==8.11.1
python-json-logger==2.0.7
websockets==12.0
httpx==0.25.2
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.1
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.0.1
